Basic structuring
Add type definitions for an expression and the the curvature types these
expressions may have. Define the structure of the rules being applied to
expressions, and define a function to apply an arbitrary rule passed as
an argument, which, by now, does nothing more than simply calling it.
Make a sample rule, walking the expression tree in post-order, always
returning UNKNOWN, and make a simple unit tests to check whether it
works properly and yields the expected result.