Split files, improve some code
Move the definitions of function properties and their corresponding
functions to a separate file named property.d. Remove the unnecessary
@property keyword. Add wrapper functions to be also able to check
for properties given a Property tuple itself to simplify the code.
Rename conv-test.d to ruleset.d containing the rules to check for
convexity. Move hasChildren() outside the Expression for some
flexibility. In the Makefile, move the -main flag from DFLAGS to
LDFLAGS, since otherwise a main function would be added to each file,
resulting in multiple definitions.