Fix division by linear functions
Fix a mistake where division by linear functions would again return
"linear" as a result. This occured because arithmeticRule simply
returned the complement of a Property, which for a linear function is
linear. Make arithmeticRule check for this case and adjust the result
accordingly. Implement some more functions to check whether a Property
is linear, increasing, or concave. Add some unit tests to verify the
functionality of arithmeticRule.