Separate and extend arithmetic operations checks
Split up arithmeticRule into four separate functions, each dealing with
their respective operation. Extend these functions in a way that the
result of adding or subtracting two functions can be checked for.
For that purpose, introduce new functions, "stronger" and "weaker" to
return the stronger/weaker property out of two; for example the addition
of two linear functions results in another linear functions, while the
addition of a linear and a (strictly) convex function results in a
convex one. The "stronger" function can later be used to get the best
result in case multiple rules are defined for a single expression
identifier.
Together with these changes comes a new system of classifying an
expression's identifier. The Classifier enum provides an additional
layar of abstraction between the expressions and the functions
operating on them. For instance, the rules to operate on arithmetic
expressions do not need to know the numerical value of a scalar but
simply whether it is positive or negative, and its position. They may,
to a certain degree, not even need to know whether a constant value is a
scalar, a vector, or a matrix, but simply whether that is positive or
not. This is an important step towards extending the existing program
to vectors and matrices, which will be added in the future.
Add a few more helper functions and move the existing functions to
classify an expression from expression.d to the new classifier.d.