Remove assertions from stronger/weaker functions
From the stronger/weaker functions of property.d remove the conditions that
given curvatures/gradients must be equal. Instead, handle the case of opposing
arguments in a more logical manner. In the 'stronger' function, when a convex
and a concave curvature are given, the function will be assumed to be linear. In
the 'weaker' function, it will be unspecified. The same logic applies to the
corresponding functions for gradients. This will simplify their usage as they
can now be called without first checking whether the conditions are met. Add
another function returning the stronger Property out of two, and adjust the
rules in arithmetics.d accordingly.