Change increasing to nondecreasing and vice versa
Change the definition of the Gradient enum to include the case of constant
functions and maintain consistency with the Curvature enum, where a linear
function may also be classified concave/convex. Since the criteria to check for
convexity are usually defined for nonincreasing/nondecreasing functions instead
of increasing/decreasing ones, the new definition makes more sense.