Remove arithmeticRule
Remove arithmeticRule completely and make the four functions dedicated
to each arithmetic operation independent from it. Removing
arithmeticRule results in shorter functions, improving readability and
gaining more structure, allowing the unit tests to be put in a more
logical place. To perform the preparatory work previously done by
arithmeticRule without doing it four times introduce a new function
"caller", which takes the expression as well as the rule to be called as
an argument, and wrap the four functions around it. Move the handling of
unary plus and unary minus into "analyze" itself, where it belongs
better. Also, remove the assertions that there must be exactly two
children to the passed expression and return unknownResult instead.
Returning unknownResult is a more graceful way of indicating that
certain input could not be handled instead of crashing, and will
simplify the process of applying several rules to the same expression,
where then the best available result will be taken.