Fix mistake in "weaker" functions
Fix a mistake in the "weaker" function where a linear and a
concave/convex curvature would result in "unspecified". Since linear can
be both concave or convex the result should be one of these. The same
applies to the "weaker" function for gradients with a constant and a
nonincreasing/nondecreasing argument.