Add Classifier variable to Expression
Add the classifier for each expression as an attribute to the expression
itself and classify each expression upon construction. This will again
gain some more abstraction and save the functions from having to
classify the expressions themselves. To achieve this, change the syntax
of the classifying functions and have the ones taking an expression as
argument operate on the classifier instead. Move the content of
classifier.d into expression.d to resolve mutual dependencies.