Change vector representation
Change the respresentation of vectors - a vector may either consist of one
and the same argument, or only of scalars. The former shall be classified as a function argument, the
latter either as a positive vector, a negative vector, or undefined. Since functions can no longer
be contained in a vector (a function may, however, be applied to all members of a vector), dotProduct
becomes obsolete and will be replaced by existing rules checking arithmetic expressions. Remove
toVector since it will be no longer used. Add unit tests to ensure correct classifications.