Introduce scalar multiplication
Add new multiplication symbol s*, representing multiplication with a
scalar, and extend the existing ruleset. Because of the current
classification system, the existing multiplication function works
already fine with either type, and no more change is necessary. This is
done for completeness, because we will introduce matrix multiplication,
which probably behaves slightly different than the previous two.